Rearrange for Aries Static Agent
