+++ requirements.txt
@@ -1,5 +1,4 @@
 toml
-python3-indy
 pytest
 aiohttp
 sortedcontainers
